removing Otto dependency in the library; now it's used only in unit tests
