+++ example-dagger/build.gradle
@@ -20,9 +20,10 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:22.2.1'
     compile project(':network-events-library')
+    compile 'com.squareup:otto:1.3.6'
     compile 'com.jakewharton:butterknife:6.1.0'
     compile 'com.squareup.dagger:dagger:1.2.2'
+    compile 'com.android.support:appcompat-v7:22.2.1'
     provided 'com.squareup.dagger:dagger-compiler:1.2.2'
 }
