+++ example/build.gradle
@@ -20,6 +20,7 @@ android {
 }
 
 dependencies {
-    compile 'com.android.support:appcompat-v7:22.2.1'
     compile project(':network-events-library')
+    compile 'com.squareup:otto:1.3.6'
+    compile 'com.android.support:appcompat-v7:22.2.1'
 }
