+++ network-events-library/build.gradle
@@ -30,9 +30,7 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.squareup:otto:1.3.6'
-
+    androidTestCompile 'com.squareup:otto:1.3.6' // Otto event bus is used only in unit tests now
     androidTestCompile 'com.android.support.test:testing-support-lib:0.1'
     androidTestCompile('com.google.truth:truth:0.25') {
         exclude group: 'junit' // Android has JUnit built in
