+++ network-events-library/src/androidTest/java/com/github/pwittchen/networkevents/library/receiver/InternetConnectionChangeReceiverTest.java
@@ -19,10 +19,10 @@ import android.support.test.runner.AndroidJUnit4;
 
 import com.github.pwittchen.networkevents.library.ConnectivityStatus;
 import com.github.pwittchen.networkevents.library.bus.BusWrapper;
-import com.github.pwittchen.networkevents.library.bus.OttoBusWrapper;
 import com.github.pwittchen.networkevents.library.logger.Logger;
 import com.github.pwittchen.networkevents.library.NetworkState;
-import com.github.pwittchen.networkevents.library.TestUtils;
+import com.github.pwittchen.networkevents.library.utils.OttoBusWrapper;
+import com.github.pwittchen.networkevents.library.utils.TestUtils;
 import com.github.pwittchen.networkevents.library.event.ConnectivityChanged;
 import com.squareup.otto.Bus;
 import com.squareup.otto.ThreadEnforcer;
@@ -31,7 +31,6 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mock;
 import org.mockito.Mockito;
 
 import java.util.ArrayList;
