+++ network-events-library/src/androidTest/java/com/github/pwittchen/networkevents/library/receiver/NetworkConnectionChangeReceiverTest.java
@@ -20,9 +20,9 @@ import android.support.test.runner.AndroidJUnit4;
 import com.github.pwittchen.networkevents.library.ConnectivityStatus;
 import com.github.pwittchen.networkevents.library.NetworkState;
 import com.github.pwittchen.networkevents.library.Task;
-import com.github.pwittchen.networkevents.library.TestUtils;
+import com.github.pwittchen.networkevents.library.utils.OttoBusWrapper;
+import com.github.pwittchen.networkevents.library.utils.TestUtils;
 import com.github.pwittchen.networkevents.library.bus.BusWrapper;
-import com.github.pwittchen.networkevents.library.bus.OttoBusWrapper;
 import com.github.pwittchen.networkevents.library.event.ConnectivityChanged;
 import com.github.pwittchen.networkevents.library.logger.Logger;
 import com.squareup.otto.Bus;
