+++ network-events-library/src/main/java/com/github/pwittchen/networkevents/library/receiver/NetworkConnectionChangeReceiver.java
@@ -19,11 +19,10 @@ import android.content.Context;
 import android.content.Intent;
 
 import com.github.pwittchen.networkevents.library.ConnectivityStatus;
-import com.github.pwittchen.networkevents.library.bus.BusWrapper;
-import com.github.pwittchen.networkevents.library.logger.Logger;
 import com.github.pwittchen.networkevents.library.NetworkHelper;
 import com.github.pwittchen.networkevents.library.Task;
-import com.squareup.otto.Bus;
+import com.github.pwittchen.networkevents.library.bus.BusWrapper;
+import com.github.pwittchen.networkevents.library.logger.Logger;
 
 public final class NetworkConnectionChangeReceiver extends BaseBroadcastReceiver {
