+++ network-events-library/src/main/java/com/github/pwittchen/networkevents/library/receiver/WifiSignalStrengthChangeReceiver.java
@@ -18,11 +18,10 @@ package com.github.pwittchen.networkevents.library.receiver;
 import android.content.Context;
 import android.content.Intent;
 
-import com.github.pwittchen.networkevents.library.bus.BusWrapper;
-import com.github.pwittchen.networkevents.library.logger.Logger;
 import com.github.pwittchen.networkevents.library.NetworkHelper;
+import com.github.pwittchen.networkevents.library.bus.BusWrapper;
 import com.github.pwittchen.networkevents.library.event.WifiSignalStrengthChanged;
-import com.squareup.otto.Bus;
+import com.github.pwittchen.networkevents.library.logger.Logger;
 
 public final class WifiSignalStrengthChangeReceiver extends BaseBroadcastReceiver {
